fix: rename data folder to mocks
